+++ src/test/java/com/google/api/generator/gapic/composer/grpc/ServiceSettingsClassComposerTest.java
@@ -31,8 +31,18 @@ public class ServiceSettingsClassComposerTest {
   public static Collection<Object[]> data() {
     return Arrays.asList(
         new Object[][] {
-          {"EchoSettings", TestProtoLoader.instance().parseShowcaseEcho()},
-          {"DeprecatedServiceSettings", TestProtoLoader.instance().parseDeprecatedService()}
+          {
+            "EchoSettings",
+            TestProtoLoader.instance().parseShowcaseEcho(),
+            "localhost:7469",
+            "v1beta1"
+          },
+          {
+            "DeprecatedServiceSettings",
+            TestProtoLoader.instance().parseDeprecatedService(),
+            "localhost:7469",
+            "v1"
+          }
         });
   }
 
@@ -40,6 +50,12 @@ public class ServiceSettingsClassComposerTest {
 
   @Parameterized.Parameter(1)
   public GapicContext context;
+
+  @Parameterized.Parameter(2)
+  public String apiShortNameExpected;
+
+  @Parameterized.Parameter(3)
+  public String apiVersionExpected;
 
   @Test
   public void generateServiceSettingsClasses() {
@@ -52,5 +68,7 @@ public class ServiceSettingsClassComposerTest {
         "servicesettings",
         clazz.classDefinition().packageString(),
         clazz.samples());
+    Assert.assertCodeEquals(clazz.apiShortName(), apiShortNameExpected);
+    Assert.assertCodeEquals(clazz.apiVersion(), apiVersionExpected);
   }
 }